refactor: rename validation functions
