+++ aio/aio-proxy/aio_proxy/parsers/date_parser.py
@@ -7,7 +7,7 @@ from aio_proxy.decorators.value_exception import value_exception_handler
     error="Veuillez indiquer une date sous le format : aaaa-mm-jj. Exemple : "
     "'1990-01-02'"
 )
-def parse_date(request, param: str, default_value=None):
+def parse_and_validate_date(request, param: str, default_value=None):
     """Extract date param from request.
 
     Args:
@@ -28,7 +28,7 @@ def parse_date(request, param: str, default_value=None):
 @value_exception_handler(
     error="Veuillez indiquer une date minimale inférieure à la date maximale."
 )
-def validate_dates(min_date=None, max_date=None):
+def validate_date_range(min_date=None, max_date=None):
     if min_date and max_date:
         min_date = datetime.strptime(min_date, "%Y-%m-%d")
         max_date = datetime.strptime(max_date, "%Y-%m-%d")
