+++ aio/aio-proxy/aio_proxy/response/parameters.py
@@ -2,7 +2,7 @@ from typing import Dict, Tuple, Union
 
 from aio_proxy.parsers.activite_principale import validate_activite_principale
 from aio_proxy.parsers.code_postal import validate_code_postal
-from aio_proxy.parsers.date_parser import parse_date, validate_dates
+from aio_proxy.parsers.date_parser import parse_and_validate_date, validate_date_range
 from aio_proxy.parsers.departement import validate_departement
 from aio_proxy.parsers.empty_params import check_empty_params
 from aio_proxy.parsers.entrepreneur_individuel import (
@@ -66,9 +66,11 @@ def extract_text_parameters(
     prenoms_dirigeant = parse_and_normalize_parameter(
         request, param="prenoms_dirigeant"
     )
-    min_date_naiss_dirigeant = parse_date(request, param="date_naissance_dirigeant_min")
-    max_date_naiss_dirigeant = parse_date(request, param="date_naissance_dirigeant_max")
-    validate_dates(min_date_naiss_dirigeant, max_date_naiss_dirigeant)
+    min_date_naiss_dirigeant = parse_and_validate_date(request,
+                                           param="date_naissance_dirigeant_min")
+    max_date_naiss_dirigeant = parse_and_validate_date(request,
+                                           param="date_naissance_dirigeant_max")
+    validate_date_range(min_date_naiss_dirigeant, max_date_naiss_dirigeant)
 
     parameters = {
         "terms": terms,
