+++ aio/aio-proxy/aio_proxy/tests/unit_tests/test_date_naissance_range_dirigeant.py
@@ -1,0 +1,14 @@
+import pytest
+from aio_proxy.parsers.date_parser import validate_date_range
+
+
+@pytest.mark.parametrize(
+    "min_date_naissance, max_date_naissance",
+    [("1964-05-08", "1963-06-04")],
+)
+def test_validate_dates_fail(min_date_naissance: str, max_date_naissance: str):
+    with pytest.raises(
+        ValueError,
+        match="Veuillez indiquer une date minimale inférieure à la date maximale.",
+    ):
+        validate_date_range(min_date_naissance, max_date_naissance)
